add like and dislike

+++ models/like.go
@@ -1,9 +1,9 @@
 package forum
 
 type Like struct {
-	ID   int64
-	Value int64
+	ID      int64
+	Value   int
 	User_ID int64
 	Post_ID int64
-	Date string
+	Date    string
 }
